restructured project for future expansion. changed module activation method
