+++ util/util.js
@@ -1,0 +1,10 @@
+function isIterable(obj) {
+  if (obj == null) {
+    return false;
+  }
+  return typeof obj[Symbol.iterator] === "function";
+}
+
+module.exports = {
+  isIterable,
+};
